setting: #38 인사평가 엔터티 수정 및 통합
