+++ perflow-backend/src/main/java/com/touchdown/perflowbackend/perfomance/command/domain/aggregate/HrPerfo.java
@@ -1,20 +1,22 @@
-package com.touchdown.perflowbackend.hrperfo.command.domain.aggregate;
+package com.touchdown.perflowbackend.perfomance.command.domain.aggregate;
 
 import com.touchdown.perflowbackend.employee.command.domain.aggregate.Employee;
 import jakarta.persistence.*;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import java.time.Instant;
 
+@NoArgsConstructor
 @Getter
-@Setter
 @Entity
 @Table(name = "hr_perfo", schema = "perflow")
 public class HrPerfo {
     @Id
     @Column(name = "hr_perfo_id", nullable = false)
-    private Long id;
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long hrPerfoId;
 
     @ManyToOne(fetch = FetchType.LAZY, optional = false)
     @JoinColumn(name = "emp_id", nullable = false)
