+++ perflow-backend/src/main/java/com/touchdown/perflowbackend/perfomance/command/domain/aggregate/Perfo.java
@@ -1,21 +1,22 @@
-package com.touchdown.perflowbackend.perfo.command.domain.aggregate;
+package com.touchdown.perflowbackend.perfomance.command.domain.aggregate;
 
 import com.touchdown.perflowbackend.employee.command.domain.aggregate.Employee;
-import com.touchdown.perflowbackend.perfoquestion.command.domain.aggregate.Perfoquestion;
 import jakarta.persistence.*;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import java.time.Instant;
 
+@NoArgsConstructor
 @Getter
-@Setter
 @Entity
 @Table(name = "perfo", schema = "perflow")
 public class Perfo {
     @Id
     @Column(name = "perfo_id", nullable = false)
-    private Long id;
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long perfoId;
 
     @ManyToOne(fetch = FetchType.LAZY, optional = false)
     @JoinColumn(name = "perfo_question_id", nullable = false)
